+++ egui/src/containers/panel.rs
@@ -212,7 +212,7 @@ impl SidePanel {
         panel_ui.expand_to_include_rect(panel_rect);
         let frame = frame.unwrap_or_else(|| Frame::side_top_panel(ui.style()));
         let inner_response = frame.show(&mut panel_ui, |ui| {
-            ui.set_min_height(ui.max_rect_finite().height()); // Make sure the frame fills the full height
+            ui.set_min_height(ui.max_rect().height()); // Make sure the frame fills the full height
             ui.set_min_width(*width_range.start());
             add_contents(ui)
         });
@@ -473,7 +473,7 @@ impl TopBottomPanel {
         panel_ui.expand_to_include_rect(panel_rect);
         let frame = frame.unwrap_or_else(|| Frame::side_top_panel(ui.style()));
         let inner_response = frame.show(&mut panel_ui, |ui| {
-            ui.set_min_width(ui.max_rect_finite().width()); // Make the frame fill full width
+            ui.set_min_width(ui.max_rect().width()); // Make the frame fill full width
             ui.set_min_height(*height_range.start());
             add_contents(ui)
         });
@@ -588,7 +588,7 @@ impl CentralPanel {
     ) -> InnerResponse<R> {
         let Self { frame } = self;
 
-        let panel_rect = ui.available_rect_before_wrap_finite();
+        let panel_rect = ui.available_rect_before_wrap();
         let mut panel_ui = ui.child_ui(panel_rect, Layout::top_down(Align::Min));
 
         let frame = frame.unwrap_or_else(|| Frame::central_panel(ui.style()));