+++ egui/src/grid.rs
@@ -109,11 +109,6 @@ impl GridLayout {
     }
 
     pub(crate) fn available_rect(&self, region: &Region) -> Rect {
-        // required for putting CollapsingHeader in anything but the last column:
-        self.available_rect_finite(region)
-    }
-
-    pub(crate) fn available_rect_finite(&self, region: &Region) -> Rect {
         let is_last_column = Some(self.col + 1) == self.num_columns;
 
         let width = if is_last_column {
@@ -135,7 +130,7 @@ impl GridLayout {
 
         let available = region.max_rect.intersect(region.cursor);
 
-        let height = region.max_rect_finite().max.y - available.top();
+        let height = region.max_rect.max.y - available.top();
         let height = height
             .at_least(self.min_cell_size.y)
             .at_most(self.max_cell_size.y);
@@ -351,7 +346,8 @@ impl Grid {
         // If somebody wants to wrap more things inside a cell,
         // then we should pick a default layout that matches that alignment,
         // which we do here:
-        ui.allocate_ui_at_rect(ui.cursor(), |ui| {
+        let max_rect = ui.cursor().intersect(ui.max_rect());
+        ui.allocate_ui_at_rect(max_rect, |ui| {
             ui.horizontal(|ui| {
                 let id = ui.make_persistent_id(id_source);
                 let grid = GridLayout {