+++ egui_demo_lib/src/apps/fractal_clock.rs
@@ -54,7 +54,7 @@ impl FractalClock {
         let painter = Painter::new(
             ui.ctx().clone(),
             ui.layer_id(),
-            ui.available_rect_before_wrap_finite(),
+            ui.available_rect_before_wrap(),
         );
         self.paint(&painter);
         // Make sure we allocate what we used (everything)