+++ egui_demo_lib/src/frame_history.rs
@@ -64,7 +64,7 @@ impl FrameHistory {
 
         // TODO: we should not use `slider_width` as default graph width.
         let height = ui.spacing().slider_width;
-        let size = vec2(ui.available_size_before_wrap_finite().x, height);
+        let size = vec2(ui.available_size_before_wrap().x, height);
         let (rect, response) = ui.allocate_at_least(size, Sense::hover());
         let style = ui.style().noninteractive();